+++ NEWS.rst
@@ -1,6 +1,12 @@
 News in rouver 0.5.2
 ====================
 
+API Additions
+-------------
+
+* ``rouver.handler``: Add ``RouteHandlerBase.respond_with_content()``.
+* ``rouver.response``: Add ``respond_with_content()``.
+
 Improvements
 ------------